refactor(auth): migrate from Roles to Role entity and update user ID types
- Replace Roles entity with new Role entity including description field
- Update all references from Roles to Role throughout the codebase
- Change userId type from Integer to Long in notifications and DTOs
- Remove unused Permissions, PermissionsRoles, and PermissionsRolesId entities
- Update UserCredentials, UserSecurity, and UserRole to use users entity
- Enhance schedule entity with proper annotations and validation
- Add password security fields to users entity
- Update repository interfaces to use Long for user IDs
- Remove deprecated User entity and related dependencies
- Update test files to use new Role entity structure

BREAKING CHANGE: Role entity structure has changed with new fields and type changes

+++ Backend/SGH/src/main/java/com/horarios/SGH/Controller/InAppNotificationController.java
@@ -48,7 +48,7 @@ public class InAppNotificationController {
             @RequestParam(defaultValue = "20") int size) {
 
         try {
-            Integer userId = getCurrentUserId();
+            Long userId = getCurrentUserId();
             log.info("Obteniendo notificaciones para usuario ID: {}", userId);
 
             Page<InAppNotification> notifications = inAppNotificationService.getActiveNotificationsByUserId(userId, page, size);
@@ -97,7 +97,7 @@ public class InAppNotificationController {
                description = "Marca una notificación específica como leída para el usuario autenticado")
     public ResponseEntity<?> markAsRead(@PathVariable Long notificationId) {
         try {
-            Integer userId = getCurrentUserId();
+            Long userId = getCurrentUserId();
 
             // Verificar que la notificación pertenece al usuario
             InAppNotification notification = inAppNotificationService.getNotificationById(notificationId);
@@ -145,7 +145,7 @@ public class InAppNotificationController {
                  description = "Obtiene el número de notificaciones no leídas del usuario")
     public ResponseEntity<?> getUnreadCount() {
         try {
-            Integer userId = getCurrentUserId();
+            Long userId = getCurrentUserId();
 
             long unreadCount = inAppNotificationService.getUnreadCount(userId);
 
@@ -214,7 +214,7 @@ public class InAppNotificationController {
     /**
      * Obtiene el ID del usuario actual desde el contexto de seguridad
      */
-    private Integer getCurrentUserId() {
+    private Long getCurrentUserId() {
         log.info("Iniciando obtención de userId actual");
         
         try {
